wallet manual fetch
