+++ apps/admin_panel/assets/src/omg-transaction/action.js
@@ -9,11 +9,10 @@ export const transfer = ({ fromAddress, toAddress, tokenId, amount }) => async d
       amount
     })
     if (result.data.success) {
-      dispatch({ type: 'TRANSACTION/CREATE/SUCCESS', transaction: result.data.data })
+      return dispatch({ type: 'TRANSACTION/CREATE/SUCCESS', transaction: result.data.data })
     } else {
-      dispatch({ type: 'TRANSACTION/CREATE/FAILED', error: result.data.data })
+      return dispatch({ type: 'TRANSACTION/CREATE/FAILED', error: result.data.data })
     }
-    return result.data
   } catch (error) {
     return dispatch({ type: 'TRANSACTION/CREATE/FAILED', error })
   }
@@ -30,7 +29,7 @@ export const getTransactions = ({ page, search, perPage }) => async dispatch => 
     })
     dispatch(hideLoading())
     if (result.data.success) {
-      return dispatch({ type: 'TRANSACTIONS/REQUEST/SUCCESS', transactions: result.data.data })
+      return dispatch({ type: 'TRANSACTIONS/REQUEST/SUCCESS', transactions: result.data.data.data })
     } else {
       return dispatch({ type: 'TRANSACTIONS/REQUEST/FAILED', error: result.data.data })
     }
