+++ apps/admin_panel/assets/src/omg-wallet/action.js
@@ -27,7 +27,7 @@ export const getWalletsByUserId = ({ userId, search }) => async dispatch => {
       userId
     })
     if (result.data.success) {
-      return dispatch({ type: 'USER_WALLETS/REQUEST/SUCCESS', wallets: result.data.data.data })
+      return dispatch({ type: 'USER_WALLETS/REQUEST/SUCCESS', wallets: result.data.data })
     } else {
       return dispatch({ type: 'USER_WALLETS/REQUEST/FAILED', error: result.data.data })
     }
