--- apps/admin_panel/assets/src/omg-wallet/walletsProvider.js
@@ -1,44 +1,0 @@
-import { Component } from 'react'
-import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
-import { selectWalletsByAccountId, selectWalletsLoadingStatus } from './selector'
-import { getWalletsByAccountId } from './action'
-class WalletsProvider extends Component {
-  static propTypes = {
-    render: PropTypes.func,
-    wallets: PropTypes.array,
-    getWalletsByAccountId: PropTypes.func,
-    walletsLoadingStatus: PropTypes.string,
-    search: PropTypes.string,
-    accountId: PropTypes.string
-  }
-  componentWillReceiveProps = nextProps => {
-    if (this.props.search !== nextProps.search) {
-      this.props.getWalletsByAccountId({
-        accountId: this.props.accountId,
-        search: this.props.search
-      })
-    }
-  }
-  componentDidMount = () => {
-    this.props.getWalletsByAccountId({
-      accountId: this.props.accountId,
-      search: this.props.search
-    })
-  }
-  render () {
-    return this.props.render({
-      wallets: this.props.wallets,
-      loadingStatus: this.props.walletsLoadingStatus
-    })
-  }
-}
-export default connect(
-  (state, props) => {
-    return {
-      wallets: selectWalletsByAccountId(state, props.search, props.accountId),
-      walletsLoadingStatus: selectWalletsLoadingStatus(state)
-    }
-  },
-  { getWalletsByAccountId }
-)(WalletsProvider)
